requirements.txt
backtrader==1.9.76.123
numpy==1.24.2
pandas==1.5.3
pyfolio==0.9.2+75.g4b901f6
PyMySQL==1.0.3
python_dateutil==2.8.2
sanic==23.3.0
TA_Lib==0.4.25
